move rendering bundle to amethyst_renderer and add docs
