+++ amethyst_rendy/Cargo.toml
@@ -17,6 +17,7 @@ amethyst_core = { path = "../amethyst_core", version = "0.6" }
 amethyst_derive = { path = "../amethyst_derive", version = "0.4" }
 amethyst_error = { path = "../amethyst_error", version = "0.1" }
 amethyst_window = { path = "../amethyst_window", version = "0.1" }
+amethyst_config = { path = "../amethyst_config", version = "0.10" }
 derive-new = "0.5.6"
 failure = "0.1"
 genmesh = "0.6"
