+++ amethyst_rendy/src/lib.rs
@@ -57,6 +57,7 @@ pub use rendy;
 pub mod pass;
 
 pub mod batch;
+pub mod bundle;
 pub mod camera;
 pub mod debug_drawing;
 pub mod error;
@@ -64,6 +65,7 @@ pub mod formats;
 pub mod light;
 pub mod mtl;
 pub mod pipeline;
+pub mod plugins;
 pub mod resources;
 pub mod serde_shim;
 pub mod shape;
@@ -84,12 +86,14 @@ mod render_test_bundle;
 
 #[doc(inline)]
 pub use crate::{
+    bundle::{RenderPlugin, RenderingBundle},
     camera::{ActiveCamera, Camera},
     formats::{
         mesh::MeshPrefab,
         texture::{ImageFormat, TexturePrefab},
     },
     mtl::{Material, MaterialDefaults},
+    plugins::*,
     sprite::{Sprite, SpriteRender, SpriteSheet, SpriteSheetFormat},
     system::{GraphCreator, RenderingSystem},
     transparent::Transparent,
