+++ examples/rendy/main.rs
@@ -12,8 +12,8 @@ use amethyst::{
     controls::{FlyControlBundle, FlyControlTag},
     core::{
         ecs::{
-            Component, DenseVecStorage, DispatcherBuilder, Entities, Entity, Join, Read,
-            ReadExpect, ReadStorage, Resources, System, SystemData, Write, WriteStorage,
+            Component, DenseVecStorage, Entities, Entity, Join, Read, ReadStorage, System, Write,
+            WriteStorage,
         },
         math::{Unit, UnitQuaternion, Vector3},
         Time, Transform, TransformBundle,
@@ -21,45 +21,32 @@ use amethyst::{
     gltf::GltfSceneLoaderSystem,
     input::{is_close_requested, is_key_down, Axis, Bindings, Button, InputBundle, StringBindings},
     prelude::*,
+    renderer::{
+        camera::{ActiveCamera, Camera},
+        debug_drawing::DebugLines,
+        light::{Light, PointLight},
+        mtl::{Material, MaterialDefaults},
+        palette::{LinSrgba, Srgb, Srgba},
+        rendy::{
+            mesh::{Normal, Position, Tangent, TexCoord},
+            texture::palette::load_from_linear_rgba,
+        },
+        resources::Tint,
+        shape::Shape,
+        sprite::{SpriteRender, SpriteSheet},
+        transparent::Transparent,
+        types::{DefaultBackend, Mesh, Texture},
+        visibility::BoundingSphere,
+        RenderFlat2D, RenderPbr3D, RenderSkybox, RenderToWindow, RenderingBundle, RenderDebugLines,
+    },
     utils::{
         application_root_dir,
         auto_fov::{AutoFov, AutoFovSystem},
         fps_counter::FpsCounterBundle,
         tag::TagFinder,
     },
-    window::{ScreenDimensions, Window, WindowBundle},
 };
-use amethyst_rendy::{
-    camera::{ActiveCamera, Camera},
-    debug_drawing::DebugLines,
-    light::{Light, PointLight},
-    mtl::{Material, MaterialDefaults},
-    palette::{LinSrgba, Srgb, Srgba},
-    pass::{
-        DrawFlat2DDesc, DrawFlat2DTransparentDesc, DrawPbrDesc, DrawPbrTransparentDesc,
-        DrawSkyboxDesc,
-    },
-    rendy::{
-        factory::Factory,
-        graph::{
-            render::{RenderGroupDesc, RenderPassNodeBuilder, SubpassBuilder},
-            GraphBuilder, ImageId, NodeId,
-        },
-        hal::command::{ClearDepthStencil, ClearValue},
-        mesh::{Normal, Position, Tangent, TexCoord},
-        texture::palette::load_from_linear_rgba,
-    },
-    resources::Tint,
-    shape::Shape,
-    sprite::{SpriteRender, SpriteSheet},
-    sprite_visibility::SpriteVisibilitySortingSystem,
-    system::{GraphCreator, RenderingSystem},
-    transparent::Transparent,
-    types::{Backend, DefaultBackend, Mesh, Texture},
-    visibility::{BoundingSphere, VisibilitySortingSystem},
-    Format, Kind,
-};
-use std::{collections::HashMap, path::Path};
+use std::path::Path;
 
 use prefab_data::{AnimationMarker, Scene, ScenePrefabData, SpriteAnimationId};
 
@@ -564,11 +551,11 @@ fn main() -> amethyst::Result<()> {
             "sampler_interpolation",
         ]))?
         .with_bundle(
-            RenderPipeline::<DefaultBackend>::new()
-                .with_settings(serde_json::json!({ "multisample": 4 }))
-                .with_plugin(RenderToMainWindow::from_config_path(display_config_path))
-                .with_plugin(RenderStandard3d::default())
-                .with_plugin(RenderStandard2d::default())
+            RenderingBundle::<DefaultBackend>::new()
+                .with_plugin(RenderToWindow::from_config_path(display_config_path))
+                .with_plugin(RenderPbr3D::default())
+                .with_plugin(RenderFlat2D::default())
+                .with_plugin(RenderDebugLines::default())
                 .with_plugin(RenderSkybox::with_colors(
                     Srgb::new(0.82, 0.51, 0.50),
                     Srgb::new(0.18, 0.11, 0.85),
@@ -578,748 +565,4 @@ fn main() -> amethyst::Result<()> {
     let mut game = Application::new(&resources, Example::new(), game_data)?;
     game.run();
     Ok(())
-}
-
-use amethyst_core::SystemBundle;
-use amethyst_error::{format_err, Error};
-use amethyst_rendy::system::GraphicsSettings;
-/// GRAPH CREATOR PLAYGROUND
-use std::marker::PhantomData;
-
-struct RenderPipeline<B: Backend> {
-    settings: Option<GraphicsSettings>,
-    plugins: Vec<Box<dyn RenderPlugin<B>>>,
-}
-
-impl<B: Backend> RenderPipeline<B> {
-    pub fn new() -> Self {
-        Self {
-            settings: Some(GraphicsSettings(serde_json::json!({}))),
-            plugins: Vec::new(),
-        }
-    }
-
-    pub fn set_settings(&mut self, settings: serde_json::Value) {
-        self.settings = Some(GraphicsSettings(settings));
-    }
-
-    pub fn with_settings(mut self, settings: serde_json::Value) -> Self {
-        self.set_settings(settings);
-        self
-    }
-
-    pub fn add_plugin(&mut self, plugin: impl RenderPlugin<B> + 'static) {
-        self.plugins.push(Box::new(plugin));
-    }
-
-    pub fn with_plugin(mut self, plugin: impl RenderPlugin<B> + 'static) -> Self {
-        self.add_plugin(plugin);
-        self
-    }
-
-    fn into_graph(self) -> RenderPipelineGraph<B> {
-        RenderPipelineGraph {
-            plugins: self.plugins,
-            marker: PhantomData,
-        }
-    }
-}
-
-impl<'a, 'b, B: Backend> SystemBundle<'a, 'b> for RenderPipeline<B> {
-    fn build(mut self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
-        for plugin in &mut self.plugins {
-            plugin.build(builder)?;
-        }
-
-        builder.add_thread_local(RenderingSystem::<B, _>::new(
-            self.settings.take().unwrap(),
-            self.into_graph(),
-        ));
-
-        Ok(())
-    }
-}
-
-struct RenderPipelineGraph<B: Backend> {
-    plugins: Vec<Box<dyn RenderPlugin<B>>>,
-    marker: PhantomData<B>,
-}
-
-impl<B: Backend> GraphCreator<B> for RenderPipelineGraph<B> {
-    fn rebuild(&mut self, res: &Resources) -> bool {
-        let mut rebuild = false;
-        for plugin in self.plugins.iter_mut() {
-            rebuild = plugin.rebuild(res) || rebuild;
-        }
-        rebuild
-    }
-
-    fn builder(&mut self, factory: &mut Factory<B>, res: &Resources) -> GraphBuilder<B, Resources> {
-        let mut plan = RenderPlan::new();
-        for plugin in self.plugins.iter_mut() {
-            plugin.plan(&mut plan, factory, res).unwrap();
-        }
-        plan.build().unwrap()
-    }
-}
-
-trait RenderPlugin<B: Backend> {
-    fn build<'a, 'b>(&mut self, _builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
-        Ok(())
-    }
-
-    fn rebuild(&mut self, _res: &Resources) -> bool {
-        false
-    }
-
-    fn plan(
-        &mut self,
-        plan: &mut RenderPlan<B>,
-        factory: &mut Factory<B>,
-        res: &Resources,
-    ) -> Result<(), Error>;
-}
-
-struct RenderPlan<B: Backend> {
-    targets: HashMap<Target, PassPlan<B>>,
-    roots: Vec<Target>,
-}
-
-impl<B: Backend> RenderPlan<B> {
-    fn new() -> Self {
-        Self {
-            targets: Default::default(),
-            roots: vec![],
-        }
-    }
-
-    pub fn define_pass(
-        &mut self,
-        target: Target,
-        outputs: PassPlanOutputs<B>,
-    ) -> Result<(), Error> {
-        let pass_plan = self
-            .targets
-            .entry(target)
-            .or_insert_with(|| PassPlan::new(target));
-
-        let root = outputs.colors.iter().any(|c| match c {
-            OutputColor::Surface(_, _) => true,
-            _ => false,
-        });
-
-        pass_plan.set_outputs(outputs)?;
-        if root {
-            self.roots.push(target);
-        }
-
-        Ok(())
-    }
-
-    pub fn extend_pass(
-        &mut self,
-        target: Target,
-        closure: impl FnOnce(PassPlanContext<'_, B>) -> Result<(), Error> + 'static,
-    ) {
-        let pass_plan = self
-            .targets
-            .entry(target)
-            .or_insert_with(|| PassPlan::new(target));
-        pass_plan.add_extension(Box::new(closure));
-    }
-
-    pub fn build(self) -> Result<GraphBuilder<B, Resources>, Error> {
-        let mut ctx = PlanContext {
-            targets: self.targets,
-            passes: Default::default(),
-            outputs: Default::default(),
-            graph_builder: GraphBuilder::new(),
-        };
-
-        for target in self.roots {
-            // prevent evaluation of roots that were accessed recursively
-            if !ctx.passes.contains_key(&target) {
-                ctx.evaluate_target(target)?;
-            }
-        }
-
-        Ok(ctx.graph_builder)
-    }
-}
-
-enum EvaluationState {
-    Evaluating,
-    Built(NodeId),
-}
-
-impl EvaluationState {
-    fn is_built(&self) -> bool {
-        match self {
-            EvaluationState::Built(_) => true,
-            _ => false,
-        }
-    }
-}
-
-pub struct PlanContext<B: Backend> {
-    targets: HashMap<Target, PassPlan<B>>,
-    passes: HashMap<Target, EvaluationState>,
-    outputs: HashMap<TargetImage, ImageId>,
-    graph_builder: GraphBuilder<B, Resources>,
-}
-
-impl<B: Backend> PlanContext<B> {
-    pub fn mark_evaluating(&mut self, target: Target) -> Result<(), Error> {
-        match self.passes.get(&target) {
-            None => {},
-            Some(EvaluationState::Evaluating) => return Err(format_err!("Trying to evaluate {:?} render plan that is already evaluating. Circular dependency detected.", target)),
-            // this case is not a soft runtime error, as this should never be allowed by the API.
-            Some(EvaluationState::Built(_)) => panic!("Trying to reevaluate a render plan for {:?}.", target),
-        };
-        self.passes.insert(target, EvaluationState::Evaluating);
-        Ok(())
-    }
-
-    fn evaluate_target(&mut self, target: Target) -> Result<(), Error> {
-        let target = self.targets.remove(&target);
-        target.map(|pass| pass.evaluate(self)).transpose()?;
-        Ok(())
-    }
-
-    pub fn submit_pass(
-        &mut self,
-        target: Target,
-        pass: RenderPassNodeBuilder<B, Resources>,
-    ) -> Result<(), Error> {
-        match self.passes.get(&target) {
-            None => {}
-            Some(EvaluationState::Evaluating) => {}
-            // this case is not a soft runtime error, as this should never be allowed by the API.
-            Some(EvaluationState::Built(_)) => {
-                panic!("Trying to resubmit a render pass for {:?}.", target)
-            }
-        };
-        let node = self.graph_builder.add_node(pass);
-        self.passes.insert(target, EvaluationState::Built(node));
-        Ok(())
-    }
-    pub fn get_image(&mut self, image_ref: TargetImage) -> Result<ImageId, Error> {
-        self.try_get_image(image_ref)?.ok_or_else(|| {
-            format_err!(
-                "Output image {:?} is not registered by the target.",
-                image_ref
-            )
-        })
-    }
-
-    pub fn try_get_image(&mut self, image_ref: TargetImage) -> Result<Option<ImageId>, Error> {
-
-        if !self
-            .passes
-            .get(&image_ref.target())
-            .map_or(false, |t| t.is_built())
-        {
-            self.evaluate_target(image_ref.target())?;
-        }
-        Ok(self.outputs.get(&image_ref).cloned())
-    }
-
-    pub fn register_output(&mut self, output: TargetImage, image: ImageId) -> Result<(), Error> {
-        if self.outputs.contains_key(&output) {
-            return Err(format_err!(
-                "Trying to register already registered output image {:?}",
-                output
-            ));
-        }
-        self.outputs.insert(output, image);
-        Ok(())
-    }
-
-    pub fn create_image(&mut self, options: ImageOptions) -> ImageId {
-        self.graph_builder
-            .create_image(options.kind, options.levels, options.format, options.clear)
-    }
-}
-
-pub struct PassPlanContext<'a, B: Backend> {
-    evaluation: &'a mut PassPlanEvaluation<B>,
-    plan_context: &'a mut PlanContext<B>,
-}
-
-impl<'a, B: Backend> PassPlanContext<'a, B> {
-    pub fn add(&mut self, order: impl Into<i32>, action: impl IntoAction<B>) -> Result<(), Error> {
-        self.evaluation.add(order, action)
-    }
-
-    pub fn colors(&self) -> usize {
-        self.evaluation.colors()
-    }
-
-    pub fn depth(&self) -> bool {
-        self.evaluation.depth()
-    }
-
-    pub fn get_image(&mut self, image: TargetImage) -> Result<ImageId, Error> {
-        self.plan_context.get_image(image)
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
-pub enum TargetImage {
-    Color(Target, usize),
-    Depth(Target),
-}
-
-impl TargetImage {
-    pub fn target(&self) -> Target {
-        match self {
-            TargetImage::Color(target, _) => *target,
-            TargetImage::Depth(target) => *target,
-        }
-    }
-}
-
-struct PassPlanEvaluation<B: Backend> {
-    key: Target,
-    colors: usize,
-    depth: bool,
-    actions: Vec<(i32, RenedrableAction<B>)>,
-}
-
-impl<B: Backend> PassPlanEvaluation<B> {
-    fn colors(&self) -> usize {
-        self.colors
-    }
-
-    fn depth(&self) -> bool {
-        self.depth
-    }
-
-    fn add(&mut self, order: impl Into<i32>, action: impl IntoAction<B>) -> Result<(), Error> {
-        let action = action.into();
-
-        if self.colors != action.colors() {
-            return Err(format_err!(
-                "Trying to add render action with {} colors to target {:?} that expects {} colors",
-                action.colors(),
-                self.key,
-                self.colors,
-            ));
-        }
-        if self.depth != action.depth() {
-            return Err(format_err!(
-                "Trying to add render action with depth '{}' to target {:?} that expects depth '{}'",
-                action.depth(),
-                self.key,
-                self.depth,
-            ));
-        }
-
-        self.actions.push((order.into(), action));
-        Ok(())
-    }
-}
-
-struct PassPlan<B: Backend> {
-    key: Target,
-    // inputs: Vec<InputSurface>,
-    extensions: Vec<Box<dyn FnOnce(PassPlanContext<'_, B>) -> Result<(), Error> + 'static>>,
-    outputs: Option<PassPlanOutputs<B>>,
-}
-
-struct PassPlanOutputs<B: Backend> {
-    colors: Vec<OutputColor<B>>,
-    depth: Option<ImageOptions>,
-}
-
-pub struct ImageOptions {
-    kind: amethyst::renderer::rendy::hal::image::Kind,
-    levels: amethyst::renderer::rendy::hal::image::Level,
-    format: amethyst::renderer::rendy::hal::format::Format,
-    clear: Option<amethyst::renderer::rendy::hal::command::ClearValue>,
-}
-
-pub enum OutputColor<B: Backend> {
-    Image(ImageOptions),
-    Surface(
-        amethyst::renderer::rendy::wsi::Surface<B>,
-        Option<amethyst::renderer::rendy::hal::command::ClearValue>,
-    ),
-}
-
-impl<B: Backend> PassPlan<B> {
-    fn new(key: Target) -> Self {
-        Self {
-            key,
-            extensions: vec![],
-            outputs: None,
-        }
-    }
-
-    fn set_outputs(&mut self, outputs: PassPlanOutputs<B>) -> Result<(), Error> {
-        if self.outputs.is_some() {
-            return Err(format_err!("Pass {:?} already defined.", self.key));
-        }
-        self.outputs.replace(outputs);
-        Ok(())
-    }
-
-    fn add_extension(
-        &mut self,
-        extension: Box<dyn FnOnce(PassPlanContext<'_, B>) -> Result<(), Error> + 'static>,
-    ) {
-        self.extensions.push(extension);
-    }
-
-    fn evaluate(self, ctx: &mut PlanContext<B>) -> Result<(), Error> {
-        if self.outputs.is_none() {
-            return Err(format_err!(
-                "Trying to evaluate not fully defined pass {:?}. Missing `define_pass` call.",
-                self.key
-            ));
-        }
-        let mut outputs = self.outputs.unwrap();
-
-        ctx.mark_evaluating(self.key)?;
-
-        let mut evaluation = PassPlanEvaluation {
-            key: self.key,
-            actions: vec![],
-            colors: outputs.colors.len(),
-            depth: outputs.depth.is_some(),
-        };
-
-        for extension in self.extensions {
-            extension(PassPlanContext {
-                plan_context: ctx,
-                evaluation: &mut evaluation,
-            })?;
-        }
-
-        let mut subpass = SubpassBuilder::new();
-        let mut pass = RenderPassNodeBuilder::new();
-
-        evaluation.actions.sort_by_key(|a| a.0);
-        for action in evaluation.actions.drain(..).map(|a| a.1) {
-            match action {
-                RenedrableAction::RenderGroup(group) => {
-                    subpass.add_dyn_group(group);
-                }
-            }
-        }
-
-        for (i, color) in outputs.colors.drain(..).enumerate() {
-            match color {
-                OutputColor::Surface(surface, clear) => {
-                    subpass.add_color_surface();
-                    pass.add_surface(surface, clear);
-                }
-                OutputColor::Image(opts) => {
-                    let node = ctx.create_image(opts);
-                    ctx.register_output(TargetImage::Color(self.key, i), node)?;
-                    subpass.add_color(node);
-                }
-            }
-        }
-
-        if let Some(opts) = outputs.depth {
-            let node = ctx.create_image(opts);
-            ctx.register_output(TargetImage::Depth(self.key), node)?;
-            subpass.set_depth_stencil(node);
-        }
-
-        pass.add_subpass(subpass);
-        ctx.submit_pass(self.key, pass)?;
-        Ok(())
-    }
-}
-
-use amethyst::renderer::rendy::graph::render::RenderGroupBuilder;
-
-pub enum RenedrableAction<B: Backend> {
-    RenderGroup(Box<dyn RenderGroupBuilder<B, Resources>>),
-}
-
-impl<B: Backend> RenedrableAction<B> {
-    fn colors(&self) -> usize {
-        match self {
-            RenedrableAction::RenderGroup(g) => g.colors(),
-        }
-    }
-
-    fn depth(&self) -> bool {
-        match self {
-            RenedrableAction::RenderGroup(g) => g.depth(),
-        }
-    }
-}
-
-pub trait IntoAction<B: Backend> {
-    fn into(self) -> RenedrableAction<B>;
-}
-
-impl<B: Backend, G: RenderGroupBuilder<B, Resources> + 'static> IntoAction<B> for G {
-    fn into(self) -> RenedrableAction<B> {
-        RenedrableAction::RenderGroup(Box::new(self))
-    }
-}
-
-#[repr(i32)]
-pub enum RenderOrder {
-    BeforeOpaque = 90,
-    Opaque = 100,
-    AfterOpaque = 110,
-    BeforeTransparent = 190,
-    Transparent = 200,
-    AfterTransparent = 210,
-    LinearPostEffects = 300,
-    ToneMap = 400,
-    DisplayPostEffects = 500,
-}
-
-impl Into<i32> for RenderOrder {
-    fn into(self) -> i32 {
-        self as i32
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Target {
-    Main,
-    Custom(&'static str),
-}
-
-impl Default for Target {
-    fn default() -> Target {
-        Target::Main
-    }
-}
-
-#[derive(Default)]
-pub struct RenderToMainWindow {
-    target: Target,
-    config: Option<amethyst::window::DisplayConfig>,
-    dimensions: Option<ScreenDimensions>,
-    dirty: bool,
-}
-
-impl RenderToMainWindow {
-    pub fn from_config_path(path: impl AsRef<Path>) -> Self {
-        Self::from_config(amethyst::window::DisplayConfig::load(path))
-    }
-
-    pub fn from_config(display_config: amethyst::window::DisplayConfig) -> Self {
-        Self {
-            config: Some(display_config),
-            ..Default::default()
-        }
-    }
-    pub fn with_target(mut self, target: Target) -> Self {
-        self.target = target;
-        self
-    }
-}
-
-impl<B: Backend> RenderPlugin<B> for RenderToMainWindow {
-    fn build<'a, 'b>(&mut self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
-        if let Some(config) = self.config.take() {
-            WindowBundle::from_config(config).build(builder)?;
-        }
-
-        Ok(())
-    }
-
-    fn rebuild(&mut self, res: &Resources) -> bool {
-        let new_dimensions = res.try_fetch::<ScreenDimensions>();
-        use std::ops::Deref;
-        if self.dimensions.as_ref() != new_dimensions.as_ref().map(|d| d.deref()) {
-            self.dirty = true;
-            self.dimensions = new_dimensions.map(|d| d.clone());
-            return false;
-        }
-        return self.dirty;
-    }
-
-    fn plan(
-        &mut self,
-        plan: &mut RenderPlan<B>,
-        factory: &mut Factory<B>,
-        res: &Resources,
-    ) -> Result<(), Error> {
-        self.dirty = false;
-
-        let window = <ReadExpect<'_, Window>>::fetch(res);
-        let surface = factory.create_surface(&window);
-        let dimensions = self.dimensions.as_ref().unwrap();
-        let window_kind = Kind::D2(dimensions.width() as u32, dimensions.height() as u32, 1, 1);
-
-        let depth_options = ImageOptions {
-            kind: window_kind,
-            levels: 1,
-            format: Format::D32Sfloat,
-            clear: Some(ClearValue::DepthStencil(ClearDepthStencil(1.0, 0))),
-        };
-
-        plan.define_pass(
-            self.target,
-            PassPlanOutputs {
-                colors: vec![OutputColor::Surface(surface, None)],
-                depth: Some(depth_options),
-            },
-        )?;
-
-        Ok(())
-    }
-}
-
-#[derive(Default)]
-pub struct RenderStandard3d {
-    target: Target,
-}
-
-impl RenderStandard3d {
-    pub fn with_target(mut self, target: Target) -> Self {
-        self.target = target;
-        self
-    }
-}
-
-impl<B: Backend> RenderPlugin<B> for RenderStandard3d {
-    fn build<'a, 'b>(&mut self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
-        builder.add(
-            VisibilitySortingSystem::new(),
-            "visibility_system",
-            &["transform_system"],
-        );
-        Ok(())
-    }
-    fn plan(
-        &mut self,
-        plan: &mut RenderPlan<B>,
-        _factory: &mut Factory<B>,
-        _res: &Resources,
-    ) -> Result<(), Error> {
-        plan.extend_pass(self.target, |mut ctx| {
-            ctx.add(RenderOrder::Opaque, DrawPbrDesc::new().builder())?;
-            ctx.add(
-                RenderOrder::Transparent,
-                DrawPbrTransparentDesc::new().builder(),
-            )?;
-            Ok(())
-        });
-        Ok(())
-    }
-}
-
-#[derive(Default)]
-pub struct RenderStandard2d {
-    target: Target,
-}
-
-impl RenderStandard2d {
-    pub fn with_target(mut self, target: Target) -> Self {
-        self.target = target;
-        self
-    }
-}
-
-impl<B: Backend> RenderPlugin<B> for RenderStandard2d {
-    fn build<'a, 'b>(&mut self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
-        builder.add(
-            SpriteVisibilitySortingSystem::new(),
-            "sprite_visibility_system",
-            &["transform_system"],
-        );
-        Ok(())
-    }
-
-    fn plan(
-        &mut self,
-        plan: &mut RenderPlan<B>,
-        _factory: &mut Factory<B>,
-        _res: &Resources,
-    ) -> Result<(), Error> {
-        plan.extend_pass(self.target, |mut ctx| {
-            ctx.add(RenderOrder::Opaque, DrawFlat2DDesc::new().builder())?;
-            ctx.add(
-                RenderOrder::Transparent,
-                DrawFlat2DTransparentDesc::new().builder(),
-            )?;
-            Ok(())
-        });
-        Ok(())
-    }
-}
-
-#[derive(Default)]
-pub struct RenderDebugLines {
-    target: Target,
-}
-
-impl RenderDebugLines {
-    pub fn with_target(mut self, target: Target) -> Self {
-        self.target = target;
-        self
-    }
-}
-
-impl<B: Backend> RenderPlugin<B> for RenderDebugLines {
-    fn plan(
-        &mut self,
-        plan: &mut RenderPlan<B>,
-        _factory: &mut Factory<B>,
-        _res: &Resources,
-    ) -> Result<(), Error> {
-        plan.extend_pass(self.target, |mut ctx| {
-            ctx.add(RenderOrder::AfterOpaque, DrawFlat2DDesc::new().builder())?;
-            ctx.add(
-                RenderOrder::Transparent,
-                DrawFlat2DTransparentDesc::new().builder(),
-            )?;
-            Ok(())
-        });
-        Ok(())
-    }
-}
-
-#[derive(Default)]
-pub struct RenderSkybox {
-    target: Target,
-    colors: Option<(Srgb, Srgb)>,
-}
-
-impl RenderSkybox {
-    pub fn with_colors(nadir_color: Srgb, zenith_color: Srgb) -> Self {
-        Self {
-            target: Default::default(),
-            colors: Some((nadir_color, zenith_color)),
-        }
-    }
-
-    pub fn with_target(mut self, target: Target) -> Self {
-        self.target = target;
-        self
-    }
-}
-
-impl<B: Backend> RenderPlugin<B> for RenderSkybox {
-    fn plan(
-        &mut self,
-        plan: &mut RenderPlan<B>,
-        _factory: &mut Factory<B>,
-        _res: &Resources,
-    ) -> Result<(), Error> {
-        let colors = self.colors.clone();
-        plan.extend_pass(self.target, move |mut ctx| {
-            let group = if let Some((nadir, zenith)) = colors {
-                DrawSkyboxDesc::with_colors(nadir, zenith).builder()
-            } else {
-                DrawSkyboxDesc::new().builder()
-            };
-
-            ctx.add(RenderOrder::AfterOpaque, group)?;
-            Ok(())
-        });
-        Ok(())
-    }
 }
